+++ doc/CHANGELOG.md
@@ -8,6 +8,17 @@
 
 ### Changed
 
+## [0.5.8] 5 September 2026
+
+### Added
+- Added themes that are stored in local storage.
+- Added Name and Icon (finally).
+- Added 5 themes.
+
+### Fixed
+
+### Changed
+
 ## [0.5.7] 20 August 2026
 
 ### Added